+++ files/gitlab-cookbooks/gitlab-ee/attributes/default.rb
@@ -57,6 +57,7 @@ default['gitlab']['geo-secondary']['db_sslmode'] = nil
 default['gitlab']['geo-secondary']['db_sslrootcert'] = nil
 default['gitlab']['geo-secondary']['db_sslca'] = nil
 default['gitlab']['geo-secondary']['db_fdw'] = nil
+default['gitlab']['geo-secondary']['db_schema_search_path'] = nil
 
 ###
 # Geo: PostgreSQL (Tracking database)