+++ files/gitlab-cookbooks/package/libraries/config/roles/geo_secondary.rb
@@ -20,6 +20,9 @@ module GeoSecondaryRole
 
     Services.enable_group('geo')
     Gitlab['geo_secondary']['enable'] = true
+    # Rails and FDW doesn't work with pgbouncer unless the default schema for
+    # the foreign table is explicitly mentioned
+    Gitlab['geo_secondary']['db_schema_search_path'] = "'$user', gitlab_secondary, public" if Gitlab['geo_secondary']['db_fdw']
     Gitlab['postgresql']['wal_level'] = 'hot_standby'
     Gitlab['postgresql']['max_wal_senders'] ||= 10
     Gitlab['postgresql']['wal_keep_segments'] ||= 10